refactoring schedule job
