+++ .env.template
@@ -85,14 +85,14 @@ SQS_CRON="0/15 * * ? * *"
 # Max delay per device when the backend has to
 # call a device for schedule. This has to be able
 # to be parsed in a TimeSpan object
-CALLFORSCHEDULE_MAX_DELAY="00:01:00"
+SCHEDULE_MAX_DELAY="00:01:00"
 # Maximum amount of devices per batch that backend
 # will call at once. If the threshold is reached
 # backend will reduce the amount of letters in
 # a batch and recalculate.
-CALLFORSCHEDULE_DEVICE_THRESHOLD=100
+SCHEDULE_DEVICE_THRESHOLD=100
 
 # Maximum amount of ads to include in the batch
-CALCULATESCHEDULEFOR_DEVICE_THRESHOLD=10
+SCHEDULE_AD_THRESHOLD=10
 # Expiration timespan for the presigned urls of ads
-CALCULATESCHEDULEFOR_URL_EXPIRE="01:00:00"
+SCHEDULE_URL_EXPIRE="01:00:00"
