+++ dotnet/integration-tests/ConfigurableBackendFactory.cs
@@ -82,12 +82,10 @@ public class ConfigurableBackendFactory : WebApplicationFactory<Program>, IAsync
             ["AWS_SQS_SECRET_KEY"] = "x",
             ["AWS_SQS_NONPROCESSED_QUEUE_URL"] = "/000000000000/nonprocessed",
             ["REQUIRE_KNOWN_DEVICES"] = "false",
-            ["CALLFORSCHEDULE_MAX_DELAY"] = "00:01:00",
-            ["CALLFORSCHEDULE_DEVICE_THRESHOLD"] = "100",
-            ["CALCULATESCHEDULEFOR_DEVICE_THRESHOLD"] = "10",
-            ["CALCULATESCHEDULEFOR_DEVICE_LOCALPATH"] =
-                $"{topLevel}/volumes.local/backend/schedules",
-            ["CALCULATESCHEDULEFOR_URL_EXPIRE"] = "01:00:00"
+            ["SCHEDULE_MAX_DELAY"] = "00:01:00",
+            ["SCHEDULE_DEVICE_THRESHOLD"] = "100",
+            ["SCHEDULE_AD_THRESHOLD"] = "10",
+            ["SCHEDULE_URL_EXPIRE"] = "01:00:00"
         };
 
         foreach (var kvp in backendEnvVars)
